Optimized printing of board to only print changed characters
Previously, we would reprint the entire board. On larger boards,
this would take some time and was noticable by the user due to some lag.
There was a significant delay between typing a character and viewing the
output.

This change tracks where each string should be rendered, and when
printing a new board, it finds only those string that have changed
and will only print them.